import-from-bot-1/USEFUL-ASSETS/shorts-bot/requirements.txt
discord.py>=2.3.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
click>=8.1.0
Pillow>=10.0.0
